Create README.md
This folder contains three files with raw data used for the analyses of the manuscript entitled "Tropical stingless bees exhibit positive phototaxis to the ultraviolet component of sunlight regardless of rising temperatures". The file "Ramos_Apidologie_Assay&calibrationdata" contains all the data collected for the phototactic orientation data of the 3 species of stingless bees, as well as the orientation latency data. The file "Ramos_Apidologie_Table1andFig.2_rawdata" contains the calculations for the data recorded in Table 1 of the manuscript, as well as the raw data used to make the transmittance and irradiance curves in Fig.2. The file "Ramos_Apidologie_temperaturedata_Fig3" contains the raw temperature data between the treatments and the external environment used to make Fig.3.